refactor
- add GeneralService file to list of files exempted from CodeClimate scrutiny
- refactor code to reduce duplication
