+++ .codeclimate.yml
@@ -3,3 +3,4 @@ exclude_patterns:
 - "spec/"
 - "**/models/"
 - "**/node_modules/"
+- "server/services/GeneralService"
