+++ server/helper/ArticleHelperClass.js
@@ -2,8 +2,6 @@ import ArticleService from '../services/ArticleService';
 import UserHelperClass from './UserHelperClass';
 import GeneralService from '../services/GeneralService';
 import GeneralHelperClass from './GeneralHelperClass';
-import Reply from '../models/reply';
-import Comment from '../models/comments';
 import Bookmark from '../models/bookmark';
 import Rating from '../models/rating';
 import Like from '../models/like';
@@ -29,40 +27,6 @@ class ArticleHelperClass {
     const savedArticle = await ArticleService.findById(articleId);
     if (savedArticle === null) throw new Error('Article does not exist');
     return savedArticle;
-  }
-
-  /**
- * @param {String} id - the id of the comment/reply to be validated
- * @param {String} type - type of comment - 'reply' or 'comment'
- * @returns {null} the token
- */
-  static async validateComment(id, type) {
-    try {
-      if (!id) throw new Error(`${type} Id is required`);
-      const isIdValid = GeneralHelperClass.isIdValid(id);
-      if (!isIdValid) throw new Error(`Invalid ${type}Id`);
-      const Model = type === 'comment' ? Comment : Reply;
-      const savedType = await GeneralService.findById(Model, id);
-      if (savedType === null) throw new Error(`${type} does not exist`);
-    } catch (error) {
-      throw error;
-    }
-  }
-
-  /**
-   *
-   * @param {Object} args - the argument from the graphQL schema
-   * @param {String} type - whether 'comment' or 'reply'
-   * @returns {String} - the id of a valid user
-   */
-  static async validateCommentFields(args, type) {
-    const userId = await UserHelperClass.validateUser(args.token);
-    await ArticleHelperClass.validateArticle(args.articleId);
-    await ArticleHelperClass
-      .validateComment(
-        type === 'comment' ? args.commentId : args.replyId, type
-      );
-    return userId;
   }
 
   /**
@@ -150,7 +114,7 @@ class ArticleHelperClass {
       throw new Error(`You have already liked this ${result.textPlaceholder}`);
     }
     if (type === 'unlike' && !existingRating) {
-      throw new Error('You haven\'t liked this article');
+      throw new Error(`You haven't liked this ${result.textPlaceholder}`);
     }
     if (type === 'unlike' && existingRating) {
       return existingRating._id;
