+++ server/resolvers/users.js
@@ -105,7 +105,6 @@ const UserResolver = {
     }
   },
   followUser: async (args) => {
-    args = args.followInput;
     const userId = await UserHelperClass
       .validateTwoUsers(args.token, args.userId,
         'The user you are trying to follow does not exist');
@@ -118,7 +117,6 @@ const UserResolver = {
     await FollowerService.create(followerObject);
   },
   unfollowUser: async (args) => {
-    args = args.unfollowInput;
     const userId = await UserHelperClass
       .validateTwoUsers(args.token, args.userId,
         'The user you are trying to unfollow does not exist');
